requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.0=mkl
brotli-python=1.0.9=py310h6a678d5_7
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.2.2=pyhd8ed1ab_0
cpuonly=2.0=0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.4=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.22.2=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=py310h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hff52083_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py310h5eee18b_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
packaging=24.0=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=py310h5eee18b_0
pip=23.3.1=py310h06a4308_0
platformdirs=4.2.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=2.2.2=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pyyaml=6.0.1=py310h5eee18b_0
pyzmq=25.1.2=py310h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.31.0=py310h06a4308_1
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.2.2=py310_cpu
torchvision=0.17.2=py310_cpu
tornado=6.1=py310h5764c6d_3
traitlets=5.14.2=pyhd8ed1ab_0
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py310h06a4308_1
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
